fixed OSS

+++ 11th class/MTX4/OSS1.py
@@ -1,6 +1,6 @@
 matrix = []
 for i in range(8):
-    matrix.append(list(map(int, input().split())))
+    matrix.append(list(map(float, input().split())))
 
 #prints the row number (1st is 1 not 0) where the biggest number is
